+++ glyph-brush/src/glyph_brush.rs
@@ -10,7 +10,7 @@ use std::{
     borrow::Cow,
     fmt,
     hash::{BuildHasher, BuildHasherDefault, Hash, Hasher},
-    i32,
+    i32, mem
 };
 
 /// A hash of `Section` data
@@ -57,6 +57,9 @@ pub struct GlyphBrush<'font, H = DefaultSectionHasher> {
     cache_glyph_drawing: bool,
 
     section_hasher: H,
+
+    last_pre_positioned: Vec<GlyphedSection<'font>>,
+    pre_positioned: Vec<GlyphedSection<'font>>,
 }
 
 impl<H> fmt::Debug for GlyphBrush<'_, H> {
@@ -147,6 +150,19 @@ impl<'font, H: BuildHasher> GlyphBrush<'font, H> {
         self.queue_custom_layout(section, &layout)
     }
 
+    /// Queues pre-positioned glyphs to be processed by the next call of
+    /// [`process_queued`](struct.GlyphBrush.html#method.process_queued). Can be called multiple
+    /// times.
+    pub fn queue_pre_positioned(
+        &mut self,
+        glyphs: Vec<(PositionedGlyph<'font>, Color, FontId)>,
+        bounds: Rect<f32>,
+        z: f32,
+    ) {
+        self.pre_positioned
+            .push(GlyphedSection { glyphs, bounds, z });
+    }
+
     #[inline]
     fn hash<T: Hash>(&self, hashable: &T) -> SectionHash {
         let mut s = self.section_hasher.build_hasher();
@@ -227,7 +243,9 @@ impl<'font, H: BuildHasher> GlyphBrush<'font, H> {
     {
         let current_text_state = self.hash(&(&self.section_buffer, screen_w, screen_h));
 
-        let result = if !self.cache_glyph_drawing || self.last_draw.text_state != current_text_state
+        let result = if !self.cache_glyph_drawing
+            || self.last_draw.text_state != current_text_state
+            || self.last_pre_positioned != self.pre_positioned
         {
             let mut some_text = false;
 
@@ -253,6 +271,11 @@ impl<'font, H: BuildHasher> GlyphBrush<'font, H> {
                 }
             }
 
+            for &(ref glyph, _, font_id) in self.pre_positioned.iter().flat_map(|p| &p.glyphs) {
+                self.texture_cache.queue_glyph(font_id.0, glyph.clone());
+                some_text = true;
+            }
+
             if some_text && self.texture_cache.cache_queued(update_texture).is_err() {
                 let (width, height) = self.texture_cache.dimensions();
                 return Err(BrushError::TextureTooSmall {
@@ -265,6 +288,7 @@ impl<'font, H: BuildHasher> GlyphBrush<'font, H> {
                     .section_buffer
                     .iter()
                     .map(|hash| &self.calculate_glyph_cache[hash])
+                    .chain(self.pre_positioned.iter())
                     .collect();
 
                 let mut verts = Vec::with_capacity(
@@ -364,6 +388,9 @@ impl<'font, H: BuildHasher> GlyphBrush<'font, H> {
             self.calculate_glyph_cache.clear();
             self.keep_in_cache.clear();
         }
+
+        mem::swap(&mut self.last_pre_positioned, &mut self.pre_positioned);
+        self.pre_positioned.clear();
     }
 
     /// Adds an additional font to the one(s) initially added on build.