+++ glyph-brush/src/glyph_brush/builder.rs
@@ -196,6 +196,9 @@ impl<'a, H: BuildHasher> GlyphBrushBuilder<'a, H> {
             cache_glyph_drawing: self.cache_glyph_drawing && self.cache_glyph_positioning,
 
             section_hasher: self.section_hasher,
+
+            last_pre_positioned: <_>::default(),
+            pre_positioned: <_>::default(),
         }
     }
 }