+++ gfx-glyph/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 * Optimise vertex updating by declaring 'Dynamic' usage & using explicit update calls.
+* Add `GlyphBrush::queue_pre_positioned` and example *pre_positioned*.
+* Add `SectionGeometry`, `GlyphPositioner` to glyph_brush re-exported types.
 
 # 0.13.1
 * Add `GlyphBrush::keep_cached`.