+++ glyph-brush/CHANGELOG.md
@@ -15,6 +15,7 @@
 * Implement `PartialEq` for `*Section`s
 * Implement `Clone`, `PartialEq` for `BrushError`
 * Implement `Debug`, `Clone` for other public things.
+* Add `GlyphBrush::queue_pre_positioned` for fully custom glyph positioning.
 
 # 0.2.4
 * Add `GlyphBrush::keep_cached`.